Issue#267: Add jobs for deactivating invitations
